+++ app/constants/domain.js
@@ -1,0 +1,1 @@
+export const DOMAIN = 'http://51.15.192.116:4500/';