+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Change the home screen.
 - Add the creature screen.
+- Add the spell screen and return data from API.
 
 #v0.1.0